refactor(board): split layout helpers
