+++ tests/grid-tools.test.ts
@@ -1,8 +1,8 @@
+import { applyGridLayout } from '../src/board/grid-tools';
 import {
-  applyGridLayout,
+  calculateGrid,
   calculateGridPositions,
-} from '../src/board/grid-tools';
-import { calculateGrid } from '../src/board/grid-layout';
+} from '../src/board/grid-layout';
 import { BoardLike } from '../src/board/board';
 
 describe('grid-tools', () => {
